wificell: test_fixture: Support packet capturer.
Set up the pcap router in test_fixture and spawn capturer when
ConfigureAP if specified.

With a packet capturer, TextFixture can provide over-the-air packets log
to facilitate debugging. For advanced usage (e.g. parse pcap to
verify the messages), it is also possible for the test to capture the
packets by itself with Router object returned by Pcap() function.

BUG=chromium:1034875
TEST=`tast run $DUT wifi.SimpleConnect*` and check there are pcap files
collected in test results.

Change-Id: I3ef927f537ee2749c6c8020699c0ba13f7d6da6f