Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+2
> > 
> > (1 comment)
> > 
> > > Patch Set 2:
> > > 
> > > (1 comment)
> > > 
> > > > Patch Set 2:
> > > > 
> > > > (1 comment)
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > PTAL. Is there an easy way to import protobufs from platform2 in grpc services?
> > > > 
> > > > I believe some of the tast tests are relying on the wilco_dtc protos. I created a separate ebuild that builds and installs the proto files (https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/53f948d6c37ab9b478de86e415dffacb140fdcf7/chromeos-base/wilco-dtc-grpc-protos/wilco-dtc-grpc-protos-9999.ebuild)
> > > 
> > > That only installs the generated and build go and c++ libraries. Including needs the actual proto files to be present.
> > 
> > Hmm. Can we include them like ../../.../platform2/... ?
> 
> Even if we can, the issue is that it's still part of another repository. That's why there will be inconsistency issue.

Consistency is also a concern since code generation needs to be done manually. People updating the proto may not be aware they need to generate code in tast. Perhaps keeping is as a byte[] is the safest/sanest approach, since in only requires a simple Marshall and Unmarshall and will always be consistent.

I tried including using ../ once when starting the remote wilco tests, but got lost counting directories and could not get it to work.

Patch-set: 2