tast-tests: Extend crostini.CrashReporter to match implementation
Now that we've implemented enough of the actual crash reporting code
to know what to expect, add those expectations to the
test. Specifically, This CL expects that all filenames will start with
"vm_crash", expects to also see a process log along side the existing
expectations for metadata and minidump files, and checks that the VM
specific metadata makes sense.

BUG=chromium:703926
TEST=tast run $IP_ADDR crostini.CrashReporter.artifact

Change-Id: I05ac8e2e761fe95429f25213e2b369501347f42e