Update patch set 1
Patch Set 1:

This is mostly about stashing and restoring the new crash directory, but because all the paths to access it include the session hash in the path we now need to make requests to session_manager during setup and tear down.

This would be a bit less awkward if the setup returned some kind of test state object that got passed to the tear down so we could only make the request once, but I don't think my sed skill are up to that refactoring.

Patch-set: 1
Reviewer: Gerrit User 1305461 <1305461@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>