wificell: router: Refactor for monitor type interface.
In current structure, we don't allow any PHY reuse. However, it will be
convenient if the packet capturer can be run on the same test router.
Refactor the selectInterface and selectPhy to allow at most one managed
and one monitor interface on the same PHY.

Also, return the whole NetDev instead of iface name in selectInterface,
so that the caller can access the information of selected PHY. It'll be
useful when checking if the PHY is shared and skipping interface setup.

BUG=chromium:1034875
TEST=tast run $DUT wifi.SimpleConnect.*

Change-Id: I5198ebad8bdbec6262ccd8600653577bf1f6721d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2138859
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>