kernel.ConfigVerify: Require some kernel hardening settings.
As suggested by Kees. Some of these are not set and some of these are
only valid in some arches so add them commented to get them in the
source code and not lose track of them and then we can start updating
the test to check.

BUG=chromium:1061514
TEST=Passes on amd64-generic.

Change-Id: Ic6fe57788dc15484ff292ec08b40a0e840f2f8d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2103208
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>