Use mock consent for non-Chrome local crash tests.
Consent and failures to start chrome have been very common sources of
flakiness, so use mock consent for many tests, and split some into both
mock consent and real consent versions. We hope to be able to make the
mock versions (which use neither chrome nor the real consent system)
critical more quickly.

Note - crostini.CrashReporter failed, but this is expected as of now.
(it's not fully implemented yet).

BUG=chromium:1060344
TEST=tast run --extrauseflags chrome_internal $DUT arc.AppCrash.* \
     crash.Sender* crostini.{CrashReporter,FsCorruption} \
     platform.{CrashReporterCrash,DevCoredump,KernelWarning,\
       SelinuxViolation,ServiceFailure,SuspendFailure,UdevCrash}*

Change-Id: I2f714d31e618c5693dde77ea8181db381dc0177c