wificell: hostapd: add PcapFreqOptions
When capturing WiFi packets, we need to configure the frequency, channel
width... on the monitor interface. However, the channel width option
cannot be obtained easily from the public elements of hostapd.Config.
(e.g. it needs to be derived from Mode, HTCaps, and later VHT options)
Add a function on hostapd.Config to generate the options for SetFreq so
that we can obtain the necessary options to capture packets of the
hostapd and some processing in hostapd.Config can be shared.

BUG=chromium:1034875
TEST=fast_build.sh -T

Change-Id: Ia836470a4782f6c4cf72c0a8ee34237365247a98