Update patch set 3
Patch Set 3:

(2 comments)

> Actualy, it's not only wilco_dtc_supportd, it's also VM.

Yeah looking more closely, it looks like it just took some time either for the gRPC services inside the VM to be ready, or the VSH commands to run. That makes more sense to see variance in the startup time of the VM if we are waiting on the binaries to be ready.

Patch-set: 3
Reviewer: Gerrit User 1304442 <1304442@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>