network: iw: Extend SetFreq to support setting channel width.
Current SetFreq does not support giving channel width options like
HT40+, 80. Extend it to accept related options so that we can set
up monitor interfaces for packet capturing in remote WiFi tests in
later commits.

Extending with options instead of adding new functions with different
inputs because it is easier for other objects to pass the settings it
needs. e.g. getting the SetFreqOptions for packet capturing from a
hostapd.Config.

BUG=chromium:1034875
TEST=fast_build.sh -T
TEST=On a kohaku DUT (which supports 160MHz channel width), create a
monitor interface on DUT and use SetFreq to configure it. Then verify
channel and width in the output of "iw dev $iface info". (80+80 is not
tested as it is not supported)

Change-Id: I97d8ce1c90c17452467b24efbb073a4a56e7e83f