Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Oh, I thought setup_app.go was already committed. I only wanted to submit the `GrantAndroidPermission` part.

Yeah, it can be tricky if you depend on CLs that in the progress.

I remember that I did incorrectly rebase and got thousand of not related lines into my CLs.

So, it's worst to quickly double check that your CL contains only your changes after rebasing :)

Patch-set: 5