wificell: test_fixture: support router with non default SSH port
When WFH, we usually use reverse tunnel for workstation to connect DUTs
in local network. Pass routerTarget to host.ParseSSHOption to extract
the port.

As for the DefaultWifiRouterHost, we don't have some rules to guess its
port, and it does not help much to use the same port as DUT. Keep it
with default SSH port.

BUG=chromium:1034875
TEST=`tast run -var="router=$ROUTER_IP:$ROUTER_PORT" $DUT_IP:$DUT_PORT
wifi.SimpleConnect.80211a`

Change-Id: I5b05cdf79254313b83ffa58029ef68aeb5e5082e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2107024
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>