wificell: test_fixture: support router with non default SSH port
When WFH, we usually use reverse tunnel for workstation to connect DUTs
in local network. Pass routerTarget to host.ParseSSHOption to extract
the port.

As for the DefaultWifiRouterHost, we don't have some rules to guess its
port, and it does not help much to use the same port as DUT. Keep it
with default SSH port.

BUG=chromium:1034875
TEST=`tast run -var="router=$ROUTER_IP:$ROUTER_PORT" $DUT_IP:$DUT_PORT
wifi.SimpleConnect`

Change-Id: I5b05cdf79254313b83ffa58029ef68aeb5e5082e