Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > Is this change really necessary? Do you have usages of extra args?
> > > > 
> > > > Also, to simplify usage of such function, you can use three dots operator: https://golang.org/ref/spec#Passing_arguments_to_..._parameters
> > > 
> > > It's not necessary -- the alternative is that I can expose all the options for apitrace that individual tests like the Loop option in the next change.  However, I expect for the other usages I have in mind (--per-frame-delay, and another option which hasn't been implemented crbug.com/1041995) only one test will need these additional arguments.  Therefore, I didn't think it was useful to clutter up a general API used by many tests to support those arguments.
> > > 
> > > https://chrome-internal-review.googlesource.com/c/chromeos/platform/tast-tests-private/+/2547695 is an example test which will be eventually using such functionality.  It's not possible yet because it's waiting for Crostini debian packages to be updated to have a newer apitrace.
> > 
> > If you will have many parameters and many of them will be optional, I can suggest to use options pattern as for example it done in Chrome: https://source.chromium.org/chromiumos/chromiumos/codesearch/+/master:src/platform/tast-tests/src/chromiumos/tast/local/chrome/chrome.go;l=117?q=file:tast-tests%20chrome
> > 
> > 
> > 
> > In the caller, it will look smth like:
> > 
> > trace.RunExtendedTest(ctx, s, pre.Container, "crostini_trace_g1.trace.bz2", "g1", trace.WithLoop(310, 323, 100),trace.WithFoo("foo"))).
> > 
> > WDYT?
> 
> For examples: https://github.com/tmrts/go-patterns/blob/master/idiom/functional-options.md

I think it is a bunch of unnecessary boilerplate at this time given to be what I expect to be the usages of it.  If the expectation was that the same option would be likely used in multiple tests (like the FrameLoop option), then trying to optimize for re-use makes sense, but at this point it just feels like premature generalization.

Instead of just adding a single string as an argument, we need to add an internal structure, a function to modify the structure, and additional code to convert that structure back into a string for every single single-use optional argument.  What was one line of code, has ballooned to probably around ten lines in four places instead of one, across two tast repos instead of one.  (Granted there's this upfront cost of handling the generic extra arguments).

I'll make the change since you're the owner in this area of code, but I think it's make work resulting in worse maintainability given then expected usages.

Patch-set: 2
Reviewer: Gerrit User 1108834 <1108834@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0