tast/webrtc: add first GetDisplayMedia (screen/window/tab capture) test
This CL adds the test and JS logic to exercise the getDisplayMedia() JS API
that allows users to capture the screen contents or a window or a tab of
their choice. The test relies on a Chrome flag to bypass the UI picker,
and for the purpose of this CL this is forced to be the screen, which is
then verified not to be all black.

BUG=chromium:1063449
TEST=tast -verbose run $IP webrtc.GetDisplayMedia.monitor
 on atlas, grunt, scarlet and kukui

Change-Id: Ie5557756f80ee49bd1b5e1f0cf02275fc48cd8ef