tast/webrtc: add a GetDisplayMedia (screen/window/tab capture) test
This CL adds the test and JS logic to exercise the getDisplayMedia() JS
API that allows users to capture the contents of the screen / a window
/ a tab of their choice. The test relies on a Chrome flag to bypass the
UI picker, and for the purpose of this CL this is forced to be the
screen, which is then verified not to be all black. This CL only lands
the "full screen" capture case, waiting for crrev.com/c/2113410 to roll.

BUG=chromium:1063449
TEST=tast -verbose run $IP webrtc.GetDisplayMedia.monitor
 on atlas, grunt, scarlet and kukui

Change-Id: Ie5557756f80ee49bd1b5e1f0cf02275fc48cd8ef