Fix ui.LauncherPageSwitchPerf.tablet_mode for krane and others
As described in the bug, the drag-up starting point was too below,
and sometimes it may fall into the hotseat instead of the apps-grid.
In that case, it fails to cause the page switch and therefore
no histogram data are recorded. This CL is essentially changing
the starting point slightly.

Also changed some log message, so that future debuggers can guess
more easily when the problem has happened.

BUG=chromium:1063196
TEST=tast run ui.HotseatScrollperf ui.LauncherPageSwitchPerf.tablet_mode

Change-Id: I99741a5175f8aa107387181357fdb43f4ff0a8ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2111638
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>