UserFilesGuest: Perform check in the correct namespace.
Guest sessions are mounted in a non-root mount namespace, so the test
needs to check in that same namespace, so that it can find the files.

Make the test robust by only checking in the non-root namespace if the
non-root namespace functionality is being used.

BUG=chromium:1056294
TEST=Passes on eve.

Change-Id: I7b18ec08445b3fff051be9ed31db2cddee05ee0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2113295
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>