CrashReporterCrash: Use seconds for sleep duration
It's not clear that this is a problem, but the 'm' suffix is a GNU
extension, so use a number of seconds for greater portability.

Also, extend sleep period to 5 minutes. (We're killing it at the end of
the test anyway, so the increased duration doesn't matter.)

BUG=chromium:1040266
TEST=tast run --extrauseflags chrome_internal $DUT platform.CrashReporterCrash*

Change-Id: I9935cab9b331cfd142512ee6be9fc0d58e78fbb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2116638
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Joon Ahn <joonbug@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>