tast-test: Add NetworkVPNConnect
The Tast test is migrated from network_VPNConnect.py, an Autotest.
This test is still flaky and has two major issues:
1- Sometimes the VPN fails to connect.
2- The Ping command fails with the following error message: command not
found.
The goal is to migrate the L2TP/IPSec server and make sure it's
at least as stable as the autotest version, then continue  with
migrating the OpenVPN server.

BUG=b:152080743
TEST=run tast network.NetworkVPNConnect

Change-Id: Iefd6131f10f4ae65469b034041919e52ad6b7e14