tast/video: reduce video resolution for Play.av1*
Most of the Play.* test cases use a 320x240 file, whereas the .av1
cases use a 720p; this is a wee taxing for older devices (BYT like
candy for example). This CL uses a smaller resolution video for av1,
concretely a 640x360, which should reduce the flakiness quite a bit
and also saves some download space (ToT 1.5MB vs 365KB).

BUG=chromium:1063168
TEST=tast run -verbose IP video.Play.av1* on atlas

Change-Id: I6a5316398b0d088ac48b05c4d9846c155d10ef3f