tast-tests: Make crostini.CrashReporter use s.Fatal for fatal errors
Currently crostini.CrashReporter uses s.Error when the expected crash
files are not found. This should be a fatal error, as the rest of the
test assumes these files were found. Currently we get a somewhat
uninformative panic instead of an error message when this fails.

BUG=chromium:703926
TEST=none

Change-Id: I77b55536d4f481465109e55cf521a74b81edcffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2118072
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>