tast-tests: Make sure the VK is hidden before starting the test.
arc.ChromeVirtualKeyboard.focusChangeTest assumes that the virtual
keyboard is hidden when starting the test, however, especially on slow
devices, the test can start while the VK is shown by the previous test.
Hopefully, this CL reduces the flakiness of the test.
Also this CL changes a visibility check to WaitUntilHidden. It may also
make the test more robust.

BUG=b:151902993
TEST=Ran arc.ChromeVirtualKeyboard 10 times

Change-Id: Iceaf630ea07279a5c386e5c81931cd48212756b2