tast-tests: crostini.CrashReporter: Handle non-test images
When testing this test, I ran it using test builds of the VM
image. This was incorrect because the actual testing infrastructure
uses base images, so this test will currently fail incorrectly.

First, we currently use a mock consent file which causes crash
reporter to ignore the actual metrics consent state. But because the
mock file is not accessible from the VM, the in-VM crash_reporter will
not produce crash reports. Fortunately, the crash test fixture already
has a way to enable real metrics consent, so use that instead.

Secondly, we currently trigger crashes using python, but python is not
installed in the base image. Instead change this to use a bash command
line which sends SIGABRT to itself.

BUG=chromium:703926
TEST=tast run $IP_ADDR crostini.CrashReporter.download

Change-Id: If682a4825df1df2bcf5f93fb98031997ee4afa0c