tast-tests: crostini.CrashReporter: Handle non-test images
When testing this test, I ran it using test builds of the VM
image. This was incorrect because the actual testing infrastructure
uses base images, so this test will currently fail incorrectly.

This is because we currently trigger crashes using python, but python
is not installed in the base image. Instead change this to use a bash
command line which sends SIGABRT to itself.

BUG=chromium:703926
TEST=tast run $IP_ADDR crostini.CrashReporter.download

Change-Id: If682a4825df1df2bcf5f93fb98031997ee4afa0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2120694
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>