Servo: Control consts should not start with 'Ctrl'
https://crrev.com/c/2124826 introduces several servo controls which
should reasonably start with 'Ctrl', such as 'CtrlEnter', which presses
the Ctrl and Enter keys on the keyboard.

If we follow the convention established by the existing StringControl
constants, this would lead to 'CtrlEnter' being the name of the control
which presses just Enter, and 'CtrlCtrlEnter' being the name of the
control which presses Ctrl and Enter together. That would be confusing.

The servo control names are clear enough without the Ctrl prefix.
Simplest just to remove the prefix.

BUG=chromium:993062
TEST=example.ServoEcho

Change-Id: I7d57108ee63b87fc7da5d97b93621fd6c50da401
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2125107
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>