crash: Wait for boot_collector in remote tests.
The boot collector does not always run within 15s of boot on all
devices, leading to flakes in remote tests. Wait for it to indicate that
it has finished before we check for files, and give it plenty of time to
do so.

Also increase the timeout for the remote tests, to allow more time for
the boot collector to run without timing out the whole test.

BUG=chromium:1040286,chromium:1039381,chromium:1039378
TEST=tast run $DUT crash.{EarlyCrash*,KernelCrash*}

Change-Id: I0d56016782f25916fb8ac53562d754579c3255af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2124817
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>