Update patch set 3
Patch Set 3:

(3 comments)

> Patch Set 3:
> 
> FYI 
> I think the process is to first get LGTM from within the team (niwa?), and then send out to tast owners.

It is recommended, specifically if the test scenario is complicated and need the background knowledge a lot, or if it needs some discussion to get to the agreement from project perspective.
Or, it is also ok to send a CL to both team reviewer and tast reviewer in parallel to reduce the review latency. Specifically, if the change is simpler, or if the Tast reviewer also have some knowledge/background about it.

Patch-set: 3
Reviewer: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>