Create firmware.Helper
Modeled after the hwsec.Helper, the firmware.Helper contains pointers to
a lot of features used by firmware tests.

This will decrease the setup/teardown required for individual tests. It
will also mean that we don't need to pass a Context, a Config, a Servo,
and an RPC Client every time we want to mode-switch the DUT. Instead,
we can just pass the Helper and a Context. Ditto for other low-level
operations.

BUG=chromium:993062
TEST=firmware.CheckBootMode

Change-Id: I6756941c42662cbdb88339d5d980ace0659d147f