Update patch set 10
Patch Set 10:

to be pefectly clear, "choosing" a tool with limitation is not a justification of compromised coverage, TBH. If tpm1.2 also needs such a test , eventually we need a version for TPM1.2 too and that version is going to cover TPM1.2 and TPM2.0. then why this version for TPM2.0 only?

Patch-set: 10