tast-tests/webrtc: add RTCPeerConnection VP8 simulcast test case
This CL adds a new webrtc.RTCPeerConnection test case to verify the
VP8 simulcast functionality, which is used e.g. on Meet/Hangouts.
For that:
- a new test param is added, which is wired through to the html file,
- the JS code is modified in a few points to use addTransceiver()
 and to fix the offer/answer SDPs as per simulcast needs and
- we pull two extra files from the Web Platform Tests (WPT)
upstream repo, each in its own subfolder and with associated LICENSE
file and README.

In this test case we don't care if it's implemented in hw or in sw
(see the next CL for that).

BUG=chromium:1067715
TEST=tast run $IP webrtc.RTCPeerConnection* on atlas

Change-Id: I09aa0cd59f1f5ab6da4b9b8b3bfbfa36029a2a0c