Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

This seems fine, but if we're concerned about unwanted options in the conf file, could we instead stash it at the beginning of the test, overwrite it with a known-good value, and then restore it after the test? That way we don't have to worry about options other developers leave over (which seems like it would be a huge source of flakiness in general, even after this particular issue is resolved).

Patch-set: 1
Reviewer: Gerrit User 1305461 <1305461@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1