userfiles: Skip one test while identifying flake.
security.UserFilesGuest is still flaky:
https://goto.google.com/security_userfilesguest

The source of the flake appears to be the elided check below, so remove
this check temporarily until I can identify the flake better. This way
we can confirm the theory that this is the source of the flake.

This does relax the test for security.UserFilesLoggedIn, but it's not a
huge deal: the test is failing with stricter (700) permissions than
what the test requires (710).

BUG=chromium:1056294
TEST=tast run <DUT> security.UserFilesGuest
TEST=for i in {1..10}; do <above command>; done
TEST=All executions pass.
TEST=tast run <DUT> security.UserFilesLoggedIn passes.

Change-Id: I63a295ecaa913f6d8df670b021c9ce6b54ceed00