tast-tests/webrtc: simplify MediaRecorder
ToT MediaRecorder test opens a new tab for every test (see e.g.
https://imgur.com/a/I8mIQ40) which is taxing for small devices
and unnecessary since every MediaRecorder object should be fully
independent: this CL moves them all to run back to back in a
single tab, and also especifies reduces the fake video capture
device (a.k.a Pacman) video resolution (if unspecified, this is
likely to be 640x480, which is unnecessarily large).

BUG=chromium:1065434
TEST=tast run DUT webrtc.MediaRecorder.* on krane, samus, atlas

Change-Id: I9220cf623084b8ce1d8b6a327c61721c8a0fd90a