tast-tests/video: move common video/encode structs to media/
This CL moves the logic for encoding tests from the video/encode
to a shared location in media/encoding, and adapts the necessary
callsites in video/ and arc/video*.

More precisely it etracts StreamParams and TestOption structs
to said location, sadly causing a flurry of small package name
fixes.

No functional changes intended.

BUG=chromium:1066463
TEST=tast -verbose run IP video.Play.vp9* on atlas

Change-Id: Iddbec8453f8a553f47c0556d9914078a77181243