Add list of fw-testing-config datafile names
When a Tast test is initialized, it needs to explicitly declare the name
of each datafile used. There is currently no way to declare an entire
directory of datafiles.

So, for fw-testing-configs, which uses a different config file for each
platform, we need to explicitly declare the name of each config file.

This is the approach recommended by nya@ here:
https://groups.google.com/a/chromium.org/g/tast-users/c/xpbMQGnbzUM

BUG=b:151832522
TEST=firmware.CheckBootMode
TEST=firmware.CheckBootMode but instead try loading an invalid platform

Change-Id: Ib028567c3c25d46a800955c1e04c213facb63daa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2138131
Commit-Queue: Greg Edelston <gredelston@google.com>
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>