Prefix metrics with resolution
Different devices have different cameras with different maximum
resolutions. When comparing/reporting performance numbers, we are
interesting not only in power and FPS but also in the resolution.

This CL also adds Android intents to the camera testing app for querying
supported resolutions (for recording, snapshot, preview), which is
useful for manual debugging. Furthermore, this CL increases the warmup
time before measuring power during camera preview/recording to 30 seconds.

BUG=b:153391970
TEST=tast run DUT arc.PowerCameraRecordingPerf.vm

Change-Id: I01378290c281001cef22784cc9e60dec5cbad084