Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > The tests are not working in ARCVM R boards: https://screenshot.googleplex.com/ihLmFZqBmL8 - looks like errors during decoding. Anyone has any idea?
> 
> I think it's because b/153120636. Let me add stevensd@.

These tests specifically use the hardware decoder, so they won't work on ARCVM until the virito-video device is checked in. For R specifically, I tested the virtio-video stack yesterday, but it fails somewhere in the Android side of things (tfiga's v4l2 test app works fine). It's not clear to me whether or not we expect C2VDAComponent to work on R, nor how much effort to put into fixing it if we do (given that it's being rewritten).

+akahuang

Patch-set: 1
CC: Gerrit User 1123360 <1123360@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>