tast: touchscreen: clean up abs configuration
As the last argument of createVirtual now supports the configuration of
absolute axes, let's use that for creating a virtual touchscreen device.

As a side effect, the MT_SLOT issue in the TODO is now fixed. (Probably
because evdev has two interfaces to configure absolute axes and
createVirtual tries new one first?)

BUG=None
TEST=tast run eve-DUT arc.WindowOutsideDisplay (w/ VirtualTouchscreen
manually enabled)

Change-Id: If8f6e047de1296afadd62552897ed9adbbf9cdb0