Update patch set 14
Patch Set 14:

(1 comment)

Added:
Before starting the test, the device is first rebooted to ensure that the test works fine even if a user was already logged into the device. Reboot would first logout the user - which is essential since we cannot set the boot lockbox flag if the user had previously logged in.

Patch-set: 14