tast-tests: move arc-specific DecodeAccelVDVDA(Perf) to arc/video
This CL moves the tests
video.DecodeAccelVDVDA.*     --> arc.VideoDecodeAcelVDVDA.*
video.DecodeAccelVDVDAPerf.* --> arc.VideoDecodeAcceVDVDAlPerf.*

The two files used for manipulating a command line binary are
migrated wholesale to local/media/decoding, except for the use
of testing.State that had to be removed (presubmit check),
returning an error instead of doing s.Fatal() directly etc.
Callsites are updated.

No functional changes intended.

Note that I had to add Timeout and "arc" as a dependency to
the arc/ files due to a testcheck.go run.

BUG=chromium:1066463
TEST=tast -verbose run IP arc.DecodeAccelVD* video.DecodeAccel.*
  on kohaku

Change-Id: I01bef64bce49332f33b8ba811383ef5d51f0c0d7