Crash tests: Stash the clobber crash directory.
We need to stash all directories, so also stash the clobber crash directory.

Sometimes, crashes from early boot go into
/mnt/stateful_partition/unencrypted/preserve/crash/.

We're not stashing this directory, so the crashes show up there during
crash tests. For the most part this isn't a problem, but for the
crash.Sender test, it causes us to send an extra crash and the test
fails.

While doing this, reduce code duplication for crash stashing/restoring.

BUG=chromium:1069291
TEST=../platform/tast/fast_build.sh -T
TEST=put crash in /mnt/..., tast run $DUT crash.Sender.mock_consent

Change-Id: I31eee553eb687b50afc54a8cbc73d6327e518bb5