Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > good catch.
> > 
> > Yury: WDYT? should we update the documentation mentioning that it returns milli-joules?
> > or do as the patch says?
> 
> it says "returns energy value, in joules units,"
> So, it is converted to joules....
> Or I misunderstand your Q?

Yes, correct. 
I was afraid that if we apply this CL, then the current power-tests might be affected by this change, and all of a sudden all tests will start reporting less consumption.

So the alternative way to fix it is, instead of changing the code, we change the documentation.

Patch-set: 1
Reviewer: Gerrit User 1202079 <1202079@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>