chrome: Fix GAIA login with/without OOBE
Even if |keepState| is specified, chrome://oobe/gaia-signin might not
exist when the device is just flashed with a new image without keeping
the old state.

This CL tweaks the condition for calling Oobe.showAddUserForTesting() to
make it work with/without the OOBE flow.

BUG=b:153673123
TEST=Pass dev.RemoteDesktop with/without OOBE.

Change-Id: I0aee497ab0967e2fb0554b0ed5c628c984720543
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2142661
Tested-by: Shik Chen <shik@chromium.org>
Commit-Queue: Shik Chen <shik@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>