wificell: Add preconditions for TestFixture.
The TestFixture setup takes a few seconds to finish. As we're going to
having more testcases, move it to preconditions to save testing time.

Tested with my local setup, this can save 1~2 seconds per test (mainly
from NewTestFixture). It's not much but might be more with slower SSH
and can save minutes in total when all wifi tests use it. (e.g.
SimpleConnect has a dozen of testcases)

As we have already moved some tests to wifi_matfunc and replace the
Autotest counterpart, apply the precondition to the unstable tests
first to see if the precondition is stable enough. The tests selected to
use Precondition in this cl:
- BeaconInterval
- CSA
- DuplicateBSSID
- Prefer5Ghz
- PTK
- RandomMACAddress
- SecChange

BUG=chromium:1034875
TEST=Run the tests above altogether in one `tast run` and check:
  - the tests with the same precontion are grouped together
  - logs in gRPC is properly printed
  - router log is properly kept in every tests

Change-Id: I987cbaa6cd0fc1467b0cf1e9ccdbddcafd2e1500