security.UserFilesGuest: Make check more robust.
The test was assuming that Guest sessions would always be mounted in a
non-root mount namespace. Well, that functionality was recently
reverted =).

Make the test more robust by always checking in the namespace of the
'cryptohome-namespace-mounter' process, which is guaranteed to contain
the mounts (since that is the process that performs the mounts.)

BUG=chromium:1069501
TEST=Passes with and without non-root mount namespaces for Guest sessions.

Change-Id: Iefcb1bf0e76fe7fe36d3f4cfff0e436f40967095