tast-tests: Change ARC boot timeout from 120s to 180s
This is a temporary workaround for enabling virtio-fs for ARCVM.
When the filesystem is in use, ARCVM takes more time to start up
especially when the host side file system cache is not fully
populated. We'll revert this once virtio-fs is fully optimized
later in Q2.

BUG=b:141891239
TEST=arc.Boot.vm
TEST=../platform/tast/fast_build.sh -T

Change-Id: I825f38b422a20ba30bfe48634eba66ee124de18f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2146036
Tested-by: Yusuke Sato <yusukes@chromium.org>
Auto-Submit: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yury Khmel <khmel@google.com>