network.ConfigureServiceForUserProfile*: new test
This ensures we
(a) establish a 2nd profile when logging in;
(b) configure (not connect to) an arbitrary WiFi service; and
(c) reload that service if Shill restarts for some reason.

Do this for both Guest and Normal login. Step (a) catches the regression
in crbug.com/1069067 (where login failed to "push" a 2nd profile), and
the rest is gravy.

BUG=chromium:1069067
TEST=`tast run ${HOST} network.ConfigureServiceForUserProfile*`,
     with and without https://crrev.com/c/2142612

Cq-Depend: chromium:2147474
Change-Id: Idecec794caa38116e2fe37dfe8d43df40d242062
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2153507
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>