Update patch set 19
Patch Set 19:

(5 comments)

hirokisato@ ptal
I've needed to adjust the logic in a11yTree, as I noticed that there was some flakiness in getDesktopTree (sometimes the old tree would be returned), as such I changed the heading name for EditText activity, and made it to poll until we find that tree.
I think that the flakiness is due to the automation api not being able to retrieve the tree fast enough, after the activity has been updated.

Patch-set: 19