tast-tests/webrtc: split MediaRecorderPerf into sw and hw
ToT MediaRecorderPerf reports both data when using first hw
and then sw; if the hw accelerator is not used, this is
not considered an error.

This CL splits MRPerf test cases into hw and sw, and enforces
the use of the appropriate encoder implementation.

BUG=chromium:1064996, chromium:1065434, chromium:1062707
TEST=tast run DUT webrtc.MediaRecorder.* webrtc.MediaRecorderPerf.*
  on kohaku

Change-Id: I987348d6929f75c8f20e98f9ceda0f9d92774f58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2147114
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>