Reform TabSwitchCUJ not to report aggregated InputLatency data
Through the discussion with analysts, I've realized that it was
a bad idea to aggregate the tab-switching data into input latency
category since their scale would be significantly different.

With this CL, the aggregated categories can't be customized for
the jank criteria, and latency for tab-switching would be separated,
so that its data aren't aggregated.

BUG=chromium:1069944
TEST=tast run ui.TabSwitchCUJ ui.TaskSwitchCUJ'*'

Change-Id: I3ed91e402e904c026eb51186c0b9501fb50cb16c