[Tast] Potential fixing for the flaky hotseat test
ui.HotseatScrollPerf is flaky now. One possible reason is that when the
device is switched from clamshell to tablet, we do not wait for the
animation triggered by the state switch to finish and it interferes the
gesture swipe on the hotseat. This CL adds another cpu.WaitUntilIdle
after the state switch. It is a quick fixing. If the flaky issue is
fixed by this CL, we should fix the state switch method.

BUG=chromium:1064459
TEST=manually

Change-Id: Iabf6473de71939654cc1cb7e3d43b67bf512f756