tast-tests: Fix the order of setting/restoring CPU freq configs
The order in which we set and restore CPU frequency configs matters for
some configs in SetUpBenchmark().
e.g.
Changing scaling_governor can affect the values of {min, max}_perf_pct.
So, we need to store the original values for them before modifying
scaling_governor.

So, this CL makes sure that the configs are set in a fixed order and
restored in the inverse order.
Note that we stop using map to store configs because the order of
iterations over a map is non-deterministic in Go.

BUG=chromium:1063764
TEST=Run "tast run <rammus> camera.DecodeAccelJPEGPerf" and check that the
original configs are recovered when the initial scaling_governor is
"powersave" and "performance".

Change-Id: I632fa16c012646964e7684955e9b55ec8484cc1b