Update patch set 11
Patch Set 11:

Hi Victor,

Due to the CPU cache/stress routine that will take much time (around 8 seconds while the "LengthSeconds" set as 1), we thought it could use the cancellation to test it since the success of the cancellation means it started normally. Should we need to change the name as "cpu_cache_cancelled" and "cpu_stress_cancelled" even our target is testing the started and cancelled at the same time?

Thanks.

Patch-set: 11